bitnami/requirements.txt
# This is a comment, to show how #-prefixed lines are ignored.
# install the airflow_livy_operator

apache-airflow-providers-apache-livy